requirements.txt
pygame>=2.0.0
numpy>=1.19.0
PyYAML>=5.4.0
sounddevice>=0.4.0
